Remove two problematic feeds per issue #502, point (1)
Delete poorly conceived and spammy feeds "Python for Beginners" and "It's My Code" as described in issue #502 and ensuing discussion.

This change proposal does not purport to resolve issue #502 in its entirety, as there may be other problematic feeds.  Additional analysis is required on that point. Furthermore, that issue also suggests in point (2) that an attestation be required of authors that their postings are formatted in the spirit of PEP8, which this change proposal does not address.

Be that as it may, accepting this proposal will immediately improve the quality of content for those browsing the planet and, in particular, will stop the planet from misleading beginners with two low quality, un-Pythonic, spammy feeds.